test: add agent/prompt_templates tests, relocate misplaced tests, fix prompts validation
- Add 11 agent tests: invariants, max_rounds=0 with tools, extract edge
  cases, tool_call_id propagation, duration tracking
- Add 20 prompt_templates tests: field filtering, sequence coercion,
  JSON serialization, YAML edge cases, synthesis substitution
- Fix prompt_templates.py: guard prompts_data is list before iteration
  to prevent ValueError on non-list truthy values (string, dict)
- Relocate 6 misplaced test classes from test_agent.py to correct files:
  TestValidationResultFields, TestValidationBuilder -> test_results.py
  TestAgentValidationConfig -> test_config.py
  TestValidationWorkbookParsing -> test_workbook_parser.py
  TestValidationConditionEvaluator -> test_condition_evaluator.py
  TestValidationValidatorAgentMode -> test_validation.py
- Relocate 2 misplaced test classes from test_prompt_templates.py to
  new test_screening.py: TestScreeningFallbackIntegration,
  TestScreeningSkillsPlanningTemplate
