+++ tests/test_agent.py
@@ -997,314 +997,170 @@ class TestAgentResultMissingFields:
         assert restored == original
 
 
-class TestValidationResultFields:
-    def test_prompt_result_validation_defaults(self):
-        from src.orchestrator.results.result import PromptResult
-
-        result = PromptResult(sequence=1)
-        assert result.validation_passed is None
-        assert result.validation_attempts is None
-        assert result.validation_critique is None
-
-    def test_prompt_result_validation_in_to_dict_when_set(self):
-        from src.orchestrator.results.result import PromptResult
-
-        result = PromptResult(
-            sequence=1,
-            agent_mode=True,
-            validation_passed=True,
-            validation_attempts=1,
-        )
-        d = result.to_dict()
-        assert d["validation_passed"] is True
-        assert d["validation_attempts"] == 1
-        assert d["validation_critique"] is None
-
-    def test_prompt_result_validation_in_to_dict_when_unset(self):
-        from src.orchestrator.results.result import PromptResult
-
-        result = PromptResult(sequence=1, agent_mode=True)
-        d = result.to_dict()
-        assert "validation_passed" in d
-        assert d["validation_passed"] is None
-        assert d["validation_attempts"] is None
-
-    def test_prompt_result_validation_from_dict(self):
-        from src.orchestrator.results.result import PromptResult
-
-        data = {
-            "sequence": 1,
-            "validation_passed": False,
-            "validation_attempts": 3,
-            "validation_critique": "Response too short",
-        }
-        result = PromptResult.from_dict(data)
-        assert result.validation_passed is False
-        assert result.validation_attempts == 3
-        assert result.validation_critique == "Response too short"
-
-    def test_prompt_result_validation_roundtrip(self):
-        from src.orchestrator.results.result import PromptResult
-
-        original = PromptResult(
-            sequence=5,
-            agent_mode=True,
-            validation_passed=False,
-            validation_attempts=2,
-            validation_critique="Missing numeric value",
-        )
-        restored = PromptResult.from_dict(original.to_dict())
-        assert restored.validation_passed == original.validation_passed
-        assert restored.validation_attempts == original.validation_attempts
-        assert restored.validation_critique == original.validation_critique
-
-
-class TestValidationBuilder:
-    def test_with_validation_result_pass(self):
-        from src.orchestrator.results.builder import ResultBuilder
-
-        builder = ResultBuilder({"sequence": 1, "prompt": "test"})
-        builder.with_validation_result(passed=True, attempts=1)
-        result = builder.build()
-        assert result.validation_passed is True
-        assert result.validation_attempts == 1
-        assert result.validation_critique is None
-
-    def test_with_validation_result_fail(self):
-        from src.orchestrator.results.builder import ResultBuilder
-
-        builder = ResultBuilder({"sequence": 1, "prompt": "test"})
-        builder.with_validation_result(passed=False, attempts=3, critique="Must be numeric")
-        result = builder.build()
-        assert result.validation_passed is False
-        assert result.validation_attempts == 3
-        assert result.validation_critique == "Must be numeric"
-
-    def test_with_validation_result_chains(self):
-        from src.orchestrator.results.builder import ResultBuilder
-
-        builder = ResultBuilder({"sequence": 1, "prompt": "test"})
-        result = builder.with_validation_result(passed=True, attempts=1).build()
-        assert result.validation_passed is True
-
-    def test_with_validation_result_in_dict(self):
-        from src.orchestrator.results.builder import ResultBuilder
-
-        builder = ResultBuilder({"sequence": 1, "prompt": "test"})
-        builder.with_validation_result(passed=True, attempts=2)
-        d = builder.build_dict()
-        assert d["validation_passed"] is True
-        assert d["validation_attempts"] == 2
-
-
-class TestValidationConfig:
-    def test_agent_validation_config_defaults(self):
-        from src.config import AgentValidationConfig
-
-        config = AgentValidationConfig()
-        assert config.enabled is True
-        assert config.max_retries == 2
-
-    def test_agent_config_has_validation(self):
-        from src.config import AgentConfig
-
-        config = AgentConfig()
-        assert hasattr(config, "validation")
-        assert config.validation.enabled is True
-        assert config.validation.max_retries == 2
-
-
-class TestValidationWorkbookParsing:
-    def test_prompts_headers_include_validation(self):
-        from src.orchestrator.workbook_parser import WorkbookParser
-
-        parser = WorkbookParser.__new__(WorkbookParser)
-        assert "validation_prompt" in WorkbookParser.PROMPTS_HEADERS
-        assert "max_validation_retries" in WorkbookParser.PROMPTS_HEADERS
-
-    def test_results_headers_include_validation(self):
-        from src.orchestrator.workbook_parser import WorkbookParser
-
-        assert "validation_passed" in WorkbookParser.RESULTS_HEADERS
-        assert "validation_attempts" in WorkbookParser.RESULTS_HEADERS
-        assert "validation_critique" in WorkbookParser.RESULTS_HEADERS
-
-    def test_sample_prompt_spec_includes_validation_columns(self):
-        from scripts.sample_workbooks import DEFAULT_PROMPT_HEADERS, PromptSpec
-
-        prompt = PromptSpec(
-            sequence=1,
-            name="validated",
-            prompt="Use calculate to compute 9.",
-            validation_prompt="The result must be 9.",
-            max_validation_retries=2,
-        )
-
-        row = prompt.to_row()
-        assert "validation_prompt" in DEFAULT_PROMPT_HEADERS
-        assert "max_validation_retries" in DEFAULT_PROMPT_HEADERS
-        assert row["validation_prompt"] == "The result must be 9."
-        assert row["max_validation_retries"] == 2
-
-    def test_sample_prompt_spec_includes_notes(self):
-        from scripts.sample_workbooks import DEFAULT_PROMPT_HEADERS, PromptSpec
-
-        prompt = PromptSpec(
-            sequence=1,
-            name="annotated",
-            prompt="Say hello",
-            notes="Helpful note for workbook users.",
-        )
-
-        row = prompt.to_row()
-        assert "notes" in DEFAULT_PROMPT_HEADERS
-        assert row["notes"] == "Helpful note for workbook users."
-
-
-class TestValidationConditionEvaluator:
-    def test_validation_passed_true(self):
-        from src.orchestrator.condition_evaluator import ConditionEvaluator
-
-        evaluator = ConditionEvaluator(results_by_name={})
-        result = {
-            "validation_passed": True,
-            "validation_attempts": 1,
-        }
-        value = evaluator._compute_property(result, "validation_passed", True)
-        assert value is True
-
-    def test_validation_passed_false(self):
-        from src.orchestrator.condition_evaluator import ConditionEvaluator
-
-        evaluator = ConditionEvaluator(results_by_name={})
-        result = {
-            "validation_passed": False,
-            "validation_attempts": 3,
-        }
-        value = evaluator._compute_property(result, "validation_passed", False)
-        assert value is False
-
-    def test_validation_passed_none(self):
-        from src.orchestrator.condition_evaluator import ConditionEvaluator
-
-        evaluator = ConditionEvaluator(results_by_name={})
-        result = {"validation_passed": None}
-        value = evaluator._compute_property(result, "validation_passed", None)
-        assert value is None
-
-    def test_validation_attempts(self):
-        from src.orchestrator.condition_evaluator import ConditionEvaluator
-
-        evaluator = ConditionEvaluator(results_by_name={})
-        result = {"validation_attempts": 3}
-        value = evaluator._compute_property(result, "validation_attempts", 3)
-        assert value == 3
-
-    def test_validation_attempts_none(self):
-        from src.orchestrator.condition_evaluator import ConditionEvaluator
-
-        evaluator = ConditionEvaluator(results_by_name={})
-        result = {}
-        value = evaluator._compute_property(result, "validation_attempts", None)
-        assert value == 0
-
-    def test_validation_condition_in_expression(self):
-        from src.orchestrator.condition_evaluator import ConditionEvaluator
-
-        evaluator = ConditionEvaluator(
-            results_by_name={
-                "my_prompt": {
-                    "validation_passed": True,
-                    "validation_attempts": 1,
-                },
-            }
-        )
-        assert evaluator.evaluate("{{my_prompt.validation_passed}} == True")
-        assert evaluator.evaluate("{{my_prompt.validation_attempts}} == 1")
-
-    def test_validation_failed_condition(self):
-        from src.orchestrator.condition_evaluator import ConditionEvaluator
-
-        evaluator = ConditionEvaluator(
-            results_by_name={
-                "my_prompt": {
-                    "validation_passed": False,
-                    "validation_attempts": 3,
-                    "validation_critique": "Response too short",
-                },
-            }
-        )
-        result, error = evaluator.evaluate("{{my_prompt.validation_passed}} == True")
-        assert not result
-        result, error = evaluator.evaluate("{{my_prompt.validation_attempts}} >= 3")
-        assert result
-
-
-class TestValidationValidator:
-    def test_validation_prompt_without_agent_mode_warns(self):
-        from src.orchestrator.validation import OrchestratorValidator, ValidationResult
-
-        validator = OrchestratorValidator(
-            prompts=[
-                {
-                    "sequence": 1,
-                    "prompt_name": "test",
-                    "prompt": "test prompt",
-                    "agent_mode": False,
-                    "validation_prompt": "Must be a number",
-                },
-            ],
-            config={},
-        )
-        result = ValidationResult()
-        validator._validate_agent_mode(result)
-        warnings = [
-            e
-            for e in result.errors
-            if e.severity == "warning" and e.code == "VALIDATION_WITHOUT_AGENT"
+class TestAgentResultInvariants:
+    def test_tool_calls_count_equals_len_tool_calls(self):
+        calls = [
+            ToolCallRecord(round=1, tool_name="a"),
+            ToolCallRecord(round=2, tool_name="b"),
+            ToolCallRecord(round=3, tool_name="c"),
         ]
-        assert len(warnings) == 1
+        result = AgentResult(response="done", tool_calls=calls, total_rounds=3)
+        assert result.tool_calls_count == 3
+        assert result.tool_calls_count == len(result.tool_calls)
 
-    def test_max_validation_retries_out_of_range(self):
-        from src.orchestrator.validation import OrchestratorValidator, ValidationResult
-
-        validator = OrchestratorValidator(
-            prompts=[
-                {
-                    "sequence": 1,
-                    "prompt_name": "test",
-                    "prompt": "test",
-                    "agent_mode": True,
-                    "tools": ["calculate"],
-                    "validation_prompt": "Must be a number",
-                    "max_validation_retries": 20,
-                },
-            ],
-            config={},
+    def test_failed_tool_calls_returns_exact_records(self):
+        ok_tc = ToolCallRecord(round=1, tool_name="search", error=None)
+        fail_tc1 = ToolCallRecord(round=2, tool_name="calc", error="overflow")
+        fail_tc2 = ToolCallRecord(round=3, tool_name="search", error="timeout")
+        result = AgentResult(
+            response="partial",
+            tool_calls=[ok_tc, fail_tc1, fail_tc2],
         )
-        result = ValidationResult()
-        validator._validate_agent_mode(result)
-        errors = [e for e in result.errors if e.code == "AGENT_INVALID_MAX_VALIDATION_RETRIES"]
-        assert len(errors) == 1
+        failed = result.failed_tool_calls
+        assert len(failed) == 2
+        assert failed[0] is fail_tc1
+        assert failed[1] is fail_tc2
 
-    def test_max_validation_retries_valid(self):
-        from src.orchestrator.validation import OrchestratorValidator, ValidationResult
-
-        validator = OrchestratorValidator(
-            prompts=[
-                {
-                    "sequence": 1,
-                    "prompt_name": "test",
-                    "prompt": "test",
-                    "agent_mode": True,
-                    "tools": ["calculate"],
-                    "validation_prompt": "Must be a number",
-                    "max_validation_retries": 3,
-                },
-            ],
-            config={},
+    def test_from_dict_ignores_unknown_fields(self):
+        result = AgentResult.from_dict(
+            {
+                "response": "test",
+                "total_rounds": 1,
+                "total_llm_calls": 2,
+                "status": "success",
+                "unknown_field": "ignored",
+                "extra": 42,
+            }
         )
-        result = ValidationResult()
-        validator._validate_agent_mode(result)
-        errors = [e for e in result.errors if e.code == "AGENT_INVALID_MAX_VALIDATION_RETRIES"]
-        assert len(errors) == 0
+        assert result.response == "test"
+        assert result.total_rounds == 1
+        assert result.total_llm_calls == 2
+        assert not hasattr(result, "unknown_field")
+
+    def test_to_dict_contains_serialized_tool_calls(self):
+        tc = ToolCallRecord(
+            round=2,
+            tool_name="calc",
+            tool_call_id="tc_42",
+            arguments={"expr": "1+1"},
+            result="2",
+            duration_ms=3.7,
+            error="bad",
+        )
+        result = AgentResult(
+            response="final",
+            tool_calls=[tc],
+            total_rounds=2,
+            total_llm_calls=3,
+            status="failed",
+        )
+        d = result.to_dict()
+        assert d["response"] == "final"
+        assert d["status"] == "failed"
+        assert d["total_rounds"] == 2
+        assert d["total_llm_calls"] == 3
+        assert len(d["tool_calls"]) == 1
+        assert d["tool_calls"][0]["tool_name"] == "calc"
+        assert d["tool_calls"][0]["round"] == 2
+        assert d["tool_calls"][0]["arguments"] == {"expr": "1+1"}
+        assert d["tool_calls"][0]["error"] == "bad"
+
+    def test_last_tool_name_with_multiple_calls(self):
+        result = AgentResult(
+            tool_calls=[
+                ToolCallRecord(round=1, tool_name="search"),
+                ToolCallRecord(round=2, tool_name="calc"),
+                ToolCallRecord(round=3, tool_name="summarize"),
+            ],
+        )
+        assert result.last_tool_name == "summarize"
+
+
+class TestAgentLoopMaxRoundsZeroWithTools:
+    def test_max_rounds_zero_with_registered_tools_returns_max_rounds_exceeded(self):
+        client = MockClient()
+        client.queue_response("answer")
+        registry = ToolRegistry()
+        registry.register(ToolDefinition(name="search", description="Search"))
+        registry.register_executor("search", lambda args: "results")
+        loop = AgentLoop(client, registry, max_rounds=0)
+        result = loop.execute("test", tools=["search"])
+        assert result.status == "max_rounds_exceeded"
+        assert result.total_rounds == 0
+        assert result.total_llm_calls == 0
+        assert result.response == ""
+
+
+class TestAgentLoopExtractToolCalls:
+    def test_no_assistant_messages_returns_empty(self):
+        client = MockClient()
+        client.chat_history = [
+            {"role": "user", "content": "hello"},
+        ]
+        registry = ToolRegistry()
+        loop = AgentLoop(client, registry)
+        assert loop._extract_tool_calls() == []
+
+    def test_assistant_without_tool_calls_key_returns_empty(self):
+        client = MockClient()
+        client.chat_history = [
+            {"role": "user", "content": "hello"},
+            {"role": "assistant", "content": "hi there"},
+        ]
+        registry = ToolRegistry()
+        loop = AgentLoop(client, registry)
+        assert loop._extract_tool_calls() == []
+
+    def test_multiple_assistants_reads_last_one(self):
+        client = MockClient()
+        client.chat_history = [
+            {"role": "assistant", "content": "first"},
+            {"role": "user", "content": "continue"},
+            {"role": "assistant", "content": "second", "tool_calls": [{"id": "tc_last"}]},
+        ]
+        registry = ToolRegistry()
+        loop = AgentLoop(client, registry)
+        calls = loop._extract_tool_calls()
+        assert len(calls) == 1
+        assert calls[0]["id"] == "tc_last"
+
+
+class TestAgentLoopToolCallIdPropagation:
+    def test_tool_call_id_passed_to_add_tool_result(self):
+        client = MockClient()
+        client.queue_response(
+            "using tool",
+            tool_calls=[
+                {"id": "call_abc123", "function": {"name": "search", "arguments": "{}"}},
+            ],
+        )
+        client.queue_response("done", tool_calls=None)
+        registry = ToolRegistry()
+        registry.register(ToolDefinition(name="search", description="Search"))
+        registry.register_executor("search", lambda args: "found")
+        loop = AgentLoop(client, registry, max_rounds=3)
+        loop.execute("test", tools=["search"])
+        tool_msgs = [m for m in client.chat_history if m.get("role") == "tool"]
+        assert len(tool_msgs) == 1
+        assert tool_msgs[0]["tool_call_id"] == "call_abc123"
+        assert tool_msgs[0]["content"] == "found"
+
+
+class TestAgentLoopDurationTracking:
+    def test_tool_call_records_have_positive_duration(self):
+        client = MockClient()
+        client.queue_response(
+            "using tool",
+            tool_calls=[
+                {"id": "tc_1", "function": {"name": "echo", "arguments": '{"msg": "hi"}'}},
+            ],
+        )
+        client.queue_response("done", tool_calls=None)
+        registry = ToolRegistry()
+        registry.register(ToolDefinition(name="echo", description="Echo"))
+        registry.register_executor("echo", lambda args: json.dumps(args))
+        loop = AgentLoop(client, registry, max_rounds=3)
+        result = loop.execute("test", tools=["echo"])
+        assert result.tool_calls_count == 1
+        assert result.tool_calls[0].duration_ms >= 0.0
+        assert result.tool_calls[0].arguments == {"msg": "hi"}
